collector-service/requirements.txt
flask
ncclient
xmltodict
influxdb-client
pyyaml